changing to eks label
