+++ container/chess-analysis/eks/Dockerfile
